EventConnectApi/requirements.txt
asgiref==3.7.2; python_version >= '3.7'
attrs==23.2.0; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
cloudinary==1.38.0
django==5.0.2; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-filter==23.5; python_version >= '3.7'
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-spectacular==0.27.1; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
markdown==3.5.2; python_version >= '3.8'
pillow==10.2.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python-decouple==3.8
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.33.0; python_version >= '3.8'
rpds-py==0.18.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.9.0; python_version < '3.11'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.0; python_version >= '3.8'
